+++ .vscode/settings.json
@@ -5,5 +5,8 @@
     "titleize",
     "uglifyjs",
     "vuln"
-  ]
+  ],
+  "prettier.trailingComma": "none",
+  "prettier.singleQuote": true,
+  "editor.formatOnSave": false
 }
